+++ src/components/Compare/Table/Result/index.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames/bind';
-import styles from './TableCompare.module.scss';
+import styles from './Result.module.scss';
 
 const cx = classNames.bind(styles);
