+++ src/components/Compare/Table/TableCompare/index.tsx
@@ -1,7 +1,7 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable no-nested-ternary */
 /* eslint-disable consistent-return */
 /* eslint-disable array-callback-return */
+/* eslint-disable no-nested-ternary */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { useEffect } from 'react';
 import classNames from 'classnames/bind';
 import styles from './TableCompare.module.scss';
